Don't use a single main object
