+++ build.gradle.kts
@@ -9,7 +9,7 @@ plugins {
 }
 
 group = "cloud.drakon"
-version = "1.0.0"
+version = "2.0.0-SNAPSHOT"
 
 repositories {
     mavenCentral()
